+++ src/lib/posix/write.c
@@ -3,7 +3,7 @@
 PUBLIC int write(
 int fd,
 char *buffer,
-unsigned nbytes
+size_t nbytes
 ){
   return(callm1(FS, WRITE, fd, nbytes, 0, buffer, NIL_PTR, NIL_PTR));
 }
